command: Add command.h file to make importing Klipper code easier
