+++ src/generic/armcm_boot.c
@@ -8,6 +8,10 @@
 #include "autoconf.h" // CONFIG_MCU
 #include "board/internal.h" // SysTick
 #include "board/irq.h" // irq_disable
+#include "command.h" // DECL_CONSTANT_STR
+
+// Export MCU type
+DECL_CONSTANT_STR("MCU", CONFIG_MCU);
 
 // Symbols created by armcm_link.lds.S linker script
 extern uint32_t _data_start, _data_end, _data_flash;
